Fix cargo-build-bpf
With this commit it should be possible to run "cargo-build-bpf". There's
also a shell script of the same name as the Rust executable, which is:

https://github.com/solana-labs/solana/blob/master/cargo-build-bpf

Not sure how the two are related, but at least you should be able to use
either of the two commands with the CLI flags shown in the shell script.

Changes:
- Add fenix overlay to have Rust locally available
- keep the BPF sdk in the output folder